feat: add shared TransactionIcon component
